pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "cloudbeds_pms_v1_3"
description = "OpenAPI client for Cloudbeds PMS v1.3 API."
readme = "README.md"
authors = [
    { name = "Cloudbeds"}
]
license = { file = "LICENSE" }
dependencies = [
    "urllib3 >= 2.6.3, < 3.0.0",
    "python-dateutil >= 2.8.2, < 3.0.0",
    "pydantic >= 2",
    "typing-extensions >= 4.7.1",
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",

    "Topic :: Software Development :: Libraries :: Python Modules",
]

requires-python = ">=3.9"

[project.urls]
Documentation = "https://github.com/cloudbeds/cloudbeds-api-python/tree/release/v1#README"
Repository = "https://github.com/cloudbeds/cloudbeds-api-python/tree/release/v1"

[project.optional-dependencies]
testing = [
    "pytest",
    "coverage"
]

[dependency-groups]
dev = [
    "pytest",
    "coverage",
    "pyyaml",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.dynamic]
version = { file = "VERSION" }

[tool.setuptools.package-data]
"cloudbeds_pms_v1_3" = ["py.typed"]
